--- backend/portal/api/app/v1/collection_id/__init__.py
@@ -1,5 +1,0 @@
-from flask import make_response
-
-
-def delete(collection_id: str):
-    return make_response({}, 202)
